+++ .config/nvim/lua/plugins/deps.lua
@@ -1,0 +1,6 @@
+vim.pack.add({
+  -- Dependencies
+  { src = "https://github.com/nvim-lua/plenary.nvim" },
+  { src = "https://github.com/kevinhwang91/promise-async" },
+  { src = "https://github.com/nvim-tree/nvim-web-devicons" },
+})
