+++ .config/nvim/lua/plugins/tree-sitter.lua
@@ -1,42 +1,19 @@
+vim.pack.add({
+  { src = "https://github.com/windwp/nvim-autopairs" },
+  { src = "https://github.com/windwp/nvim-ts-autotag" },
+  { src = "https://github.com/JoosepAlviste/nvim-ts-context-commentstring" },
+  { src = "https://github.com/nvim-treesitter/nvim-treesitter", branch = "main" },
+  { src = "https://github.com/nvim-treesitter/nvim-treesitter-textobjects", branch = "main" },
+})
+
+require("nvim-autopairs").setup({
+  disable_filetype = { "codecompanion" },
+})
+
+--=============================================================================
+-- Keymaps
+--=============================================================================
 local opts = { noremap = true, silent = true }
-
-require("nvim-treesitter").install({
-  "c",
-  "cmake",
-  "cpp",
-  "csv",
-  "diff",
-  "gitcommit",
-  "gitignore",
-  "go",
-  "fish",
-  "html",
-  "json",
-  "ledger",
-  "lua",
-  "markdown",
-  "markdown_inline",
-  "php",
-  "python",
-  "ruby",
-  "rust",
-  "toml",
-  "vim",
-  "vimdoc",
-  "vue",
-  "yaml",
-})
-require("nvim-treesitter").update()
-
-vim.api.nvim_create_autocmd("FileType", {
-  callback = function(args)
-    local filetype = args.match
-    local lang = vim.treesitter.language.get_lang(filetype)
-    if vim.treesitter.language.add(lang) then
-      vim.treesitter.start()
-    end
-  end,
-})
 
 om.set_keymaps("af", function()
   require("nvim-treesitter-textobjects.select").select_textobject("@function.outer", "textobjects")
@@ -54,10 +31,24 @@ om.set_keymaps("as", function()
   require("nvim-treesitter-textobjects.select").select_textobject("@local.scope", "locals")
 end, { "x", "o" }, opts)
 
-require("nvim-autopairs").setup({
-  Check_ts = true,
-  enable_moveright = true,
-  fast_wrap = {
-    map = "<c-e>",
-  },
+--=============================================================================
+-- Autocmds
+--=============================================================================
+vim.api.nvim_create_autocmd("PackChanged", {
+  callback = function()
+    vim.cmd("TSUpdate")
+  end,
+})
+
+vim.api.nvim_create_autocmd("FileType", {
+  callback = function(args)
+    local filetype = args.match
+    local lang = vim.treesitter.language.get_lang(filetype)
+    if not vim.tbl_contains(require("nvim-treesitter.config").get_available(), lang) then
+      return
+    end
+    require("nvim-treesitter").install(lang):await(function()
+      pcall(vim.treesitter.start)
+    end)
+  end,
 })
